@@ -4,9 +4,9 @@ namespace Trexima\EuropeanCvBundle\Form\Parts;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
-use Trexima\EuropeanCvBundle\Entity\EuropeanCV;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
+use Trexima\EuropeanCvBundle\Entity\EuropeanCV;
 use Trexima\EuropeanCvBundle\Form\Type\DrivingLicenseType;
 use Trexima\EuropeanCvBundle\Form\Type\EuropeanCVDrivingLicenseType;
 
@@ -17,17 +17,18 @@ use function Symfony\Component\Translation\t;
  */
 class EuropeanCVPartDrivingLicenseType extends AbstractType
 {
-
     /**
      * {@inheritdoc}
      */
-    public function buildForm(FormBuilderInterface $builder, array $options)
+    public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $entity = $builder->getData();
+
         $drivingLicenses = [];
-        foreach($entity->getDrivingLicenses() as $drivingLicense) {
+        foreach ($entity->getDrivingLicenses() as $drivingLicense) {
             $drivingLicenses[] = $drivingLicense->getDrivingLicense()->value;
         }
+
         $builder
             ->add('drivingLicenseOwner', CheckboxType::class, [
                 'required' => false,
@@ -48,25 +49,24 @@ class EuropeanCVPartDrivingLicenseType extends AbstractType
                 'entry_options' => [
                     'label' => false,
                     'attr' => [
-                        'class' => 'd-flex my-2 align-items-center gap-4'
+                        'class' => 'row row-cols-1 row-cols-sm-auto gx-4 gy-1 mb-3 align-items-center'
                     ]
                 ],
                 'required' => false,
                 'hidden' => empty($entity?->getDrivingLicenseOwner()),
                 'existing_licenses' => $drivingLicenses,
-            ])
-        ;
+            ]);
     }
 
     /**
      * {@inheritdoc}
      */
-    public function configureOptions(OptionsResolver $resolver)
+    public function configureOptions(OptionsResolver $resolver): void
     {
         parent::configureOptions($resolver);
         $resolver->setDefaults([
             'data_class' => EuropeanCV::class,
-         ]);
+        ]);
 
         $resolver->setRequired([
             'photo_upload_route'
