@@ -17,7 +17,7 @@ class EuropeanCVPhoneType extends AbstractType
     /**
      * {@inheritdoc}
      */
-    public function buildForm(FormBuilderInterface $builder, array $options)
+    public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
             ->add('prefix', EnumType::class, [
@@ -42,14 +42,13 @@ class EuropeanCVPhoneType extends AbstractType
                     'class' => 'form-control-md'
                 ],
                 'form_floating' => true
-            ])
-            ;
+            ]);
     }
 
     /**
      * {@inheritdoc}
      */
-    public function configureOptions(OptionsResolver $resolver)
+    public function configureOptions(OptionsResolver $resolver): void
     {
         parent::configureOptions($resolver);
         $resolver->setDefaults([
