@@ -5,7 +5,6 @@ namespace Trexima\EuropeanCvBundle\Form\Type;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
-use Symfony\Component\Form\Extension\Core\Type\EnumType;
 use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\Form\FormEvent;
@@ -16,22 +15,19 @@ use Symfony\Component\Validator\Constraints\Valid;
 use Symfony\Contracts\Translation\TranslatorInterface;
 use Trexima\EuropeanCvBundle\Entity\Enum\WorkBreakEnum;
 use Trexima\EuropeanCvBundle\Entity\EuropeanCVWorkBreak;
-use Trexima\EuropeanCvBundle\Form\Type\MonthYearRangeType;
 
 use function Symfony\Component\Translation\t;
 
 class EuropeanCVWorkBreakType extends AbstractType implements EventSubscriberInterface
-{   
-    public function __construct(
-        private readonly TranslatorInterface $translator
-    )
+{
+    public function __construct(private readonly TranslatorInterface $translator)
     {
     }
 
     /**
      * {@inheritdoc}
      */
-    public function buildForm(FormBuilderInterface $builder, array $options)
+    public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $entity = $builder->getData();
         $type = $entity->getType();
@@ -110,7 +106,7 @@ class EuropeanCVWorkBreakType extends AbstractType implements EventSubscriberInt
                 ]
             ])
             ->addEventSubscriber($this);
-            ;
+
             $builder->get('type')->resetViewTransformers();
     }
 
